requirements.txt
litserve
coqui-tts
torch
